# Move the components of tpc_finish into the database for MySQL