--- a/src/relstorage/adapters/mysql/__init__.py
+++ b/src/relstorage/adapters/mysql/__init__.py
@@ -51,7 +51,6 @@ from __future__ import absolute_import
 from __future__ import print_function
 
 from relstorage.adapters.mysql.adapter import MySQLAdapter
-from relstorage.adapters.mysql.adapter import select_driver
+
 
 assert MySQLAdapter
-assert select_driver