--- a/src/relstorage/adapters/mysql/adapter.py
+++ b/src/relstorage/adapters/mysql/adapter.py
@@ -57,7 +57,7 @@ from zope.interface import implementer
 from relstorage._compat import iteritems
 from relstorage.options import Options
 
-from .._abstract_drivers import _select_driver
+from ..adapter import AbstractAdapter
 
 from ..dbiter import HistoryFreeDatabaseIterator
 from ..dbiter import HistoryPreservingDatabaseIterator
@@ -79,13 +79,13 @@ from .txncontrol import MySQLTransactionControl
 
 log = logging.getLogger(__name__)
 
-def select_driver(options=None):
-    return _select_driver(options or Options(), drivers)
 
 @implementer(IRelStorageAdapter)
-class MySQLAdapter(object):
+class MySQLAdapter(AbstractAdapter):
     """MySQL adapter for RelStorage."""
     # pylint:disable=too-many-instance-attributes
+
+    driver_options = drivers
 
     def __init__(self, options=None, **params):
         if options is None:
@@ -94,7 +94,7 @@ class MySQLAdapter(object):
         self.keep_history = options.keep_history
         self._params = params
 
-        self.driver = driver = select_driver(options)
+        self.driver = driver = self._select_driver()
         log.debug("Using driver %r", driver)
 
         self.connmanager = MySQLdbConnectionManager(
@@ -180,3 +180,24 @@ class MySQLAdapter(object):
         p = sorted(iteritems(p))
         parts.extend('%s=%r' % item for item in p)
         return ", ".join(parts)
+
+    # A temporary magic variable as we move TID allocation into some
+    # databases; with an external clock, we *do* need to sleep waiting for
+    # TIDs to change in a manner we can exploit; that or we need to be very
+    # careful about choosing pack times.
+    RS_TEST_TXN_PACK_NEEDS_SLEEP = 1
+
+    def lock_database_and_choose_next_tid(self,
+                                          cursor,
+                                          username,
+                                          description,
+                                          extension):
+        proc = 'lock_and_choose_tid'
+        args = ()
+        if self.keep_history:
+            proc = proc + '(%s, %s, %s, %s)'
+            args = (False, username, description, extension)
+
+        multi_results = self.driver.callproc_multi_result(cursor, proc, args)
+        tid, = multi_results[0][0]
+        return tid