--- a/src/relstorage/adapters/mysql/procs/hf/lock_and_choose_tid.sql
+++ b/src/relstorage/adapters/mysql/procs/hf/lock_and_choose_tid.sql
@@ -1,0 +1,7 @@
+CREATE PROCEDURE lock_and_choose_tid()
+COMMENT '{CHECKSUM}'
+BEGIN
+    DECLARE next_tid_64 BIGINT;
+    CALL lock_and_choose_tid_p(next_tid_64);
+    SELECT next_tid_64;
+END;