History-free and History-preserving both, including the commit (where possible). This takes us from 7 queries down to 1, which should be a nice win for gevent or threaded environments under load.

Note that the mere act of calling the stored procedure (before it was even attempting to commit) apparently crashes MySQL 5.7.12-log as deployed on AppVeyor (once we call it we get an error and abort the transaction, and the act of aborting tries to connect to the database and fails; from then on we can no longer connect: all connections are refused). 
```
Running:
 test_abort_transaction (relstorage.tests.blob.testblob.MySQLPyMySQL_SharedHistoryFreeTestBlobTransaction) (9.778 s)


Error in test test_abort_transaction (relstorage.tests.blob.testblob.MySQLPyMySQL_SharedHistoryFreeTestBlobTransaction)
raceback (most recent call last):
...
 File "\\relstorage\tests\blob\blob_transaction.py", line 75, in test_abort_transaction
   transaction.commit()
 File "\\\transaction\_manager.py", line 252, in commit
   return self.manager.commit()
...
 File "\\\transaction\_transaction.py", line 446, in _commitResources
   self._synchronizers.map(lambda s: s.afterCompletion(self))
 File "\\\transaction\weakset.py", line 61, in map
   f(elt)
 File "\\\transaction\_transaction.py", line 446, in <lambda>
   self._synchronizers.map(lambda s: s.afterCompletion(self))
 File "\\\ZODB\Connection.py", line 757, in afterCompletion
   self.newTransaction(transaction, False)
 File "\\\ZODB\Connection.py", line 737, in newTransaction
   invalidated = self._storage.poll_invalidations()
 File "\\relstorage\storage\__init__.py", line 637, in poll_invalidations
   self.__on_load_first_use
 File "\\relstorage\adapters\connections.py", line 151, in restart_and_call
   return self.call(callback, True, *args, **kw)
 File "\\relstorage\adapters\connections.py", line 174, in call
   self._open_connection()
 File "\\relstorage\adapters\connections.py", line 114, in _open_connection
   new_conn, new_cursor = self._new_connection()
 File "\\relstorage\adapters\connmanager.py", line 221, in open_for_load
   conn, cursor = self._do_open_for_load()
 File "\\relstorage\adapters\mysql\connmanager.py", line 106, in _do_open_for_load
   replica_selector=self.ro_replica_selector)
 File "\\relstorage\adapters\mysql\connmanager.py", line 77, in open
   conn = self._db_connect(**params)
  File "\\relstorage\adapters\mysql\drivers\pymysql.py", line 81, in connect
   return AbstractMySQLDriver.connect(self, *args, **kwargs)
 File "\\relstorage\adapters\_abstract_drivers.py", line 128, in connect
   return self._connect(*args, **kwargs)
 File "\\\pymysql\__init__.py", line 94, in Connect
   return Connection(*args, **kwargs)
 File "\\\pymysql\connections.py", line 325, in __init__
   self.connect()
 File "\\\pymysql\connections.py", line 630, in connect
   raise exc
pymysql.err.OperationalError: (2003, "Can't connect to MySQL server on '127.0.0.1' ([WinError 10061] No connection could be made because the target machine actively refused it)")
```

That release is from early 2016 so I'm not *extremely* worried about it, but it would be nice to figure out eventually.