@jamadden 2019-07-26
We have a strong candidate for [the crashing bug](https://bugs.mysql.com/bug.php?id=86799), fixed in [MySQL 5.7.19 (2017-07)](https://dev.mysql.com/doc/relnotes/mysql/5.7/en/news-5-7-19.html):

> Incorrect behavior could occur for INSERT statements executed in stored-program or prepared-statement context, if the VALUES part of an ON DUPLICATE KEY UPDATE clause referred to a BLOB value in the INSERT column list. (Bug #24538207, Bug #25361251, Bug #25530880, Bug #25684790)

That's exactly what we do in a history-free database: 
```sql
  INSERT INTO object_state (
    zoid,
    tid,
    state_size,
    state
  )
  SELECT zoid,
         p_committing_tid,
         COALESCE(LENGTH(state), 0),
         state
  FROM temp_store
  ORDER BY zoid
  ON DUPLICATE KEY UPDATE
     tid = VALUES(tid),
     state_size = VALUES(state_size),
     state = VALUES(state);
```

The `state` column is stored as a `BLOB`; `p_committing_tid` is a parameter.

We hadn't seen this before because, even though the generic mover wants to prepare that query, the MySQL dialect simply won't prepare statements that require parameters. Thus the statement was never prepared before.