@jamadden 2019-07-26
Benchmark results for 25 concurrent gevent greenlets against MySQL 5.7.26 (` zodbshootout --include-mapping false --min-objects 200000 -c 25 -p 3 --loops 10 --object-counts 10 -o ../<results> --threads shared --gevent <zodb.conf> store`) appear to show some nice improvements compared to both 3.0a5 and master, which should mean overall higher throughput:

| Benchmark                            | 3.0a5 | master    | this branch   |
|--------------------------------------|------------------------|-----------------------------|------------------------------|
| mysqlclient_hp: store 10 raw pickles | 4.06 ms                | not significant             | 2.00 ms: 2.03x faster (-51%) |
| mysqlclient_hf: store 10 raw pickles | 3.43 ms                | 3.13 ms: 1.09x faster (-9%) | 2.04 ms: 1.68x faster (-41%) |